+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Added:
 
 1. Add methods:
   * *TelegramBot#removeReplyListener()* (by @githugger)
+1. Add proper error handling (by @GochoMugo)
 1. Add health-check endpoint (by @mironov)
   * `options.webHook.healthEndpoint`
 1. Use *String#indexOf()*, instead of *RegExp#test()*, to